[#32] Move bg task & notification fn to separate file
