+++ app/src/lib/index.js
@@ -1,2 +1,4 @@
 export { default as api } from './api';
 export { default as loc } from './loc';
+export { default as notification } from './notification';
+export { default as backgroundTask } from './background-task';
