+++ app/src/navigation/index.js
@@ -15,6 +15,34 @@ import {
 } from '../pages';
 import { UIState, AuthState, UserState } from '../store';
 import { BackHandler } from 'react-native';
+import * as TaskManager from 'expo-task-manager';
+import * as BackgroundFetch from 'expo-background-fetch';
+import * as Notifications from 'expo-notifications';
+import { backgroundTask, notification } from '../lib';
+
+const TASK_NAME = 'sync-form-version';
+
+Notifications.setNotificationHandler({
+  handleNotification: async () => ({
+    shouldShowAlert: true,
+    shouldPlaySound: true,
+    shouldSetBadge: true,
+  }),
+});
+
+TaskManager.defineTask(TASK_NAME, async () => {
+  try {
+    console.log('[bgTask]Function here');
+    await backgroundTask.syncFormVersion({
+      sendPushNotification: notification.sendPushNotification,
+      showNotificationOnly: true,
+    });
+    return BackgroundFetch.BackgroundFetchResult.NewData;
+  } catch (err) {
+    console.error('define task manager failed:', err);
+    return BackgroundFetch.Result.Failed;
+  }
+});
 
 const Stack = createNativeStackNavigator();
 
@@ -35,6 +63,22 @@ const RootNavigator = () => {
     });
     return () => backHandler.remove();
   }, [token, currentPage]);
+
+  React.useEffect(() => {
+    backgroundTask.backgroundTaskStatus(TASK_NAME);
+    notification.registerForPushNotificationsAsync();
+    const notificationListener = Notifications.addNotificationReceivedListener((notification) => {
+      console.log('[Notification]Received Listener');
+    });
+    const responseListener = Notifications.addNotificationResponseReceivedListener((response) => {
+      console.log('[Notification]Response Listener');
+      backgroundTask.syncFormVersion({ showNotificationOnly: false });
+    });
+    return () => {
+      Notifications.removeNotificationSubscription(notificationListener);
+      Notifications.removeNotificationSubscription(responseListener);
+    };
+  }, []);
 
   return (
     <Stack.Navigator screenOptions={{ headerShown: false }} initialRouteName={currentPage}>
